refactor: code style (#9432)
