+++ packages/plugins/Web3Profile/src/SNSAdaptor/emitter.ts
@@ -1,13 +1,13 @@
-import type { FireflyLensAccount } from '@masknet/web3-providers/types'
+import type { FireflyBaseAPI } from '@masknet/web3-providers/types'
 import { Emitter } from '@servie/events'
 import type { BadgeBounding } from './types.js'
 
 export const emitter = new Emitter<{
-    open: [{ badgeBounding: BadgeBounding; lensAccounts: FireflyLensAccount[] }]
+    open: [{ badgeBounding: BadgeBounding; lensAccounts: FireflyBaseAPI.FireflyLensAccount[] }]
     close: []
 }>()
 
-export function openPopup(badgeBounding: BadgeBounding, lensAccounts: FireflyLensAccount[]) {
+export function openPopup(badgeBounding: BadgeBounding, lensAccounts: FireflyBaseAPI.FireflyLensAccount[]) {
     emitter.emit('open', { badgeBounding, lensAccounts })
 }
