--- packages/web3-providers/src/Firefly/types.ts
@@ -1,17 +1,0 @@
-export type Result<T> = {
-    code: number
-    reason: string
-    message: string
-    data: T
-}
-
-export interface FireflyLensAccount {
-    address: string
-    name: string
-    handle: string
-    bio: string
-    url: string
-    profileUri: string[]
-}
-
-export type LensResult = Result<FireflyLensAccount[]>
