+++ packages/web3-providers/src/entry-types.ts
@@ -5,6 +5,7 @@ export * from './types/Bundler.js'
 export * from './types/DSearch.js'
 export * from './types/Domain.js'
 export * from './types/Explorer.js'
+export * from './types/Firefly.js'
 export * from './types/FungibleToken.js'
 export * from './types/Funder.js'
 export * from './types/Fuse.js'
@@ -45,4 +46,3 @@ export * from './RSS3/types.js'
 export * from './GoPlusLabs/types.js'
 
 export * from './StringStorage/index.js'
-export * from './Firefly/types.js'
