+++ packages/web3-providers/src/types/Firefly.ts
@@ -1,0 +1,23 @@
+export namespace FireflyBaseAPI {
+    export type Result<T> = {
+        code: number
+        reason: string
+        message: string
+        data: T
+    }
+
+    export interface FireflyLensAccount {
+        address: string
+        name: string
+        handle: string
+        bio: string
+        url: string
+        profileUri: string[]
+    }
+
+    export type LensResult = Result<FireflyLensAccount[]>
+
+    export interface Provider {
+        getLensByTwitterId(twitterHandle: string): Promise<FireflyLensAccount[]>
+    }
+}
